IOS-3636 Split relations builder into several entities
